requirements.txt
# Core
gemma4
transformers>=4.40.0
peft>=0.10.0
unsloth>=2024.4
bitsandbytes
scikit-learn
numpy<2.0

# RAG
chromadb>=0.4.0
sentence-transformers>=2.5.0
langchain>=0.1.0
langchain-community>=0.0.20
pypdf>=4.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Agent / Function Calling
google-generativeai>=0.3.0

# UI
gradio>=4.0.0

# Utils
tqdm
pyyaml